+++ TheWarWithin/WarlockDemonology.lua
@@ -2430,4 +2430,121 @@ spec:RegisterStateExpr( "tyrant_padding", function ()
     return gcd.max * ( settings.tyrant_padding or 1 )
 end )
 
-spec:RegisterPack( "Demonology", 20250818.1, [[Hekili:v31EZTnYr(plQszAblBkcirVYBjQR24DRAZMu(2l01LRUuhbbbaLqeianEiALsf)SF9m41my65bFzN8p2san(190VNhcyM9SppBAGxr4Sp5mYz8OBSVzOJZvFW2z20INxhoB6Ap)h9Uh(HeVvW)(ZHRstsJtV)zYTEoo1lGarEAzMpC7ztxugfx8NsMTahxBG21H(WLF)7Nn9HOGGWkAdZ9NnLq77gDZ7SV5h3o)Vg(uuEyW25RZIsZIkEE78fE0lKMSDEE0QTZlcZlIsUFiilzPlJIbj4p8h2o)HII15)4LxEFuXdLlg6NU6sG8YyVIO0e)mVLfKF3)YfXPlUS4HWnEzBasJsU8N8jK871m8VeLxKFzq4sVY4IlbQIt9F0nOvfmKGY2FJWYPT4)rc(BN)X0vRIG)F6Zj(WGjC8h(W4RDQi(NbfdCTUr72FR6g)(F7D2)yN8VzZMHzErblslYBgfWpFzw460SIldEQ8)5F8Rb5phz)Z3x8LV85Fk53wh1I0yZrAX4h)V(50)8p9X)50S0Sh(Lp)3St)YNQqA78F9N(O5qL)lF8pB)l)ZFDC6)7)4pEtXvp58BF4lbKbOhv3MpCDwi8yl8kUyYLldJDdsxfLqvDVnA5KIOvH3nAWzRdlg6fh9u4GZwuUC5W7ZIwLgLf6MU0n3ZplAzKF4WY144MxUcmsUaiem3bWiJymaFYllYBrC4BjHbtwgLLx4w8CMxc8FGe)201tYbM9KxCz4e7XheqEbb1avbqN4cQR7l9Ycgg(1q)YIWQhHO18IdtkWimHWSGJR8uRDFcc4wgfMivCeP7Okn37hmCL3xnz8)YlNsrkVCbQcIMPiYVMCoT0fnu(Gxq6g3fPX83)GKhiEQXtCKPafL4MUomjmRNRmE8160nHzU5rRFOkQLgvTjkoWnA168H5fqvJ7K8SuTczatEqOWrSlOeYcYhcjAJ8JkUZEW5lsZZhqbTrh6NMfoeuvjV8YzsUB5Aljzd6uYnzA)LkDnumj8PWmO0cvPphkgqvwPzBNhLd)4tErXeL0WwKb88(kyKazG9ARtBYGrs20ZYti7PqwY99IJDR(v3yOmtLriZZpYloxjkV8cD4Vm6(hkCZcx5fLKFRJZBlZjPZw6crfmwnPSkQiCvE)hIOz(7)A4JrXr)FBN)Fs9h2opjfQJLeggquxrqT3vP5Wvafmer1veMZttkFRDYQZ0FBJt4qbh5(WfL8u6JHUHFTimlXl2Ly)RGSYzmkzzzETjOfvWBeIcZx7Tjza2fVLglwNi5nV)cFVCwU3PlG67RJHYeevau6pYxy4IgZlvyOM3SqVaiuj(XWSAJ7GZptsIQxEXpnnMeaiMdR2lWQ7PfseY84Ijj7EET6Oone4hIf0p5dVZ5nqW3teFQH2dFWljGuM9(Y4aVelwhtPASQqBgTtoeBpOoJjpGdjIcKikXp82AcO(D8pDhrdoN2GH9ijI)p8o5Y(3xJdRMtGkIwRVi0l7rtQcO7KbTKIORAOJu6qWw1oSmNH3nXzmRxJzpfiMxThIPiH7IGs8kmJ2BNWZ3UAA(p7hdvh8YUpSiFIDN7mBLQ6qLRE351UebPPRgeewrj8Z0kDwVXzqx9XBNC1agQf4oGDErwyY9fpOS(XGQ)ZfsdVkkm)oUAfla7BUpmzU7dHYLRwuRav9i8bjeQByRxwbJ8ddx2hRx3hq2v(gpOQIAbNOmWeJlAuFT9Wh4syyEtxDw7lhBmZVhHTo6zBpD8TJ1)ihQKEdIKE9UXwcJsBkHAwCauogMnp0hfKrTVJ1vdajg8IskHeWB244E9A)b4jDRNwgxmcee7xaZdSlDCR8bTaIaDFZ0GInPU0epewb)okwCTfeZQoYkRNwQXb6mbRxJ0FAtuVYWyXgJvJ1zsaRFh6iAyMjpt7VAzDJvIsWvQd4LgKVlJiQ9MRKdBx7sKZM257Au7)U2CMtAt9js7AlZanZ825dqw6iOraAV(qF)EqkZxdxSKUYsLjfrXBNVbUnyw3o)HWmMw)hsBAEcF4zrwuYJG3OnTJ086q06RsBIi3L0Mn5UW0qBVEu(WLLjpKsUhybZ7XfH0an8XbLpos4JZEZhBi22pUmiuyebiMvU4z3npegVgSOHXXMcQJeqDoeqTDx5Lu6fJjO5RFEfKcaQJ5UjCrp9peXhMTeYF56LdZLb8M9F2npmRCLPS2bN1oQyTZXH1v(BUbLzvRDw)b)6Su)HEjp7ga5zBOckii1b8n2JT6D)8O7tawbb7qWIlDDyF(noJSmvgDukJo6KrrNxEz05yiJ2U5pN4xxGnAzZ6Eu9)UHX5HtgnC8B9ttcIi41nFUEr(dy0DTvpBguV6vAkO2q4e99G2HPcgoXCnGZbRbC6RbC(wRbqyObAGaVvEqNUndJMnAOVQWHvvyZOiotQVG4DA0rmoVXpfFht8g8R6e4tGKktqF5f5MzL2zjpwBwaR3CU9WXx0bbPYfTFGOKciTpKueibbfIxQL1DNFUYanjduzS32q2BxZE4PVGvaiwT3XBtTEL9OrwwiEFK0yKFMFD3EBECArJP1MBjJ07w1u8BW5NjFkFm6RMjIFheLDg61TgCUIfFS3sGWtfIptJN6eNbMnLI7CQ39h1dfgpVMPazg(xmMDDU68UQBmHrRu5z2GkBb8(Sw142gnxqLrZc5b75UE3eUv6DV8PCq9PCuKQAh8PCK4t58nWNY(46tzFC9PSr8PSL4tjYA1XsYteG5t5C89PO5P2L6FDjQ60usQ)oXwvWMfxJwD(ykSWst19Yl01d(wW)qw8HL2Urewle0UpnqN6GRtnjqvVo1rLZMfxRB7Qo1XyDQ93eDkTvj(PFPEjrr2rTj3mQArkyWP6io0U2lJG8Jsda60nitfuF3HitE1cDN)MaQzS(HQVVA5qEqMkbbt(TBAAu6qBhnFYNG8UV7OGPsbZPSBvuwwAgzwKlZaklZcdClsxLcxDdDjW50hkPw0Rhlg05qWXwXWzhssRpPH6eXgkfAsRPxkuN6Yy)kQmrEgxYsywz3pljekkgtAU3BDoz1JH75hMDqGkbtIAqvBqyEU2qNf7sv2AL6)6i(QCGUZSIawsRn)VIdtrp09FyA3mm7Vk3FkTEjR5pgcdRoDftuVZ6IBzF9YXJTbRdoNEVU9ZUCTqd7T30QMAXDXM8uTQbPBY9DxnQdCbQSWgRxiz3Wz3tIXVRgsYXnKCos8R3xhu4W2Q7o0Mmw2J1TDf92Td173E)tDylG47LTbsYFTBJRqVVMnWzq9IHWSnlYfYE7gZoRzu44rIbBCvRJ(iBe1EnKWWQ)zJRfmUDTQsaf2Okvco7gDzS82g7lCqOUT5O3uHIKKi9o5s6mcYfF2JQB95rBc5CaeM9yuY9soMznucmArCAAG7YYSNXUlm6cPuGDts(38csszIVVywVp)qyEy3w2vDK06LbSrp2FB6sc)ARw1pGF)0nl)C7CrKWqHfRvYrQCKYgyvDWZ(mLWTZBxw38TZxMghNUbSm0JeVF7XXR6uY3DeZLPMeL6Ut(fYPoa1L5IMZ9R0Zofw4s5Adx9MBTVc9W7nz0bpMWG9U)JEv86pkXEObcvjpLYM0cOgiJy1MpyzDuBoLAtxdM6Bzro3wdEh0uU5UlHzjv3fZEe5YTue8jci36CXC80gJULmXd0ZnHLGNRG3stw73CJrhJTwQNn9j46GMP9VCMXZMUXllbIPZNn9ZKdOmygsZkiX(zBN)AARVVMCEC)sjyxH8e5PKZYSxjmvsp6zC2hIgVpmh08)LOe4wK)it(yAcWj6TFDxVRSNIaaZIuM7wiC7ZT)QLbG2FxCragHeoWDKaUJAjw42gcQwjgLeoWVsL6O)rDaxJGqfhlUwL8RLfYOIJfJvnk4xwi8XGanMdV8LacNvkPNJTVxLItZOcNgZHFhgv6PNJT)GuLj65SarjkJUJmBiXIU7o8Adk1Ngrj4yNKejo14uYXQBK6eyMAsoDhz2O0AihETwd9PivcUrwdvuYXQpi1WlCccqm5y0CsH)Kb83j5gMhuff5v)fe9A5RjTk8pTNHimoozKvZaFpfzdfJDrI1jWDTKnsAeME7lon8T8DSX)0H83njxVRVWPNdJbNyxF8ZpNuxFZezdfJDrI1jWDMbzthP6mAHvjP6Y8UjkMudgk24OiFIg7NZ2PgFtqw8yZHankrMk7gZbPZivg27B5Ttn(MGSbAfjezQSBmh2d9(XjQzxIGD0Ko(eDYyrtik55KD2u75EB6PIfTnejJsPS2wlRXprSf9osSf8Nj2gR23qT(LF)06IS(BSw)YATEx4G8LLQDhKWsvXCp(WRVv4jE6buvjRJKtj6MGl7X9us5lMBFkL2tl6D4kFvjL7r4OWJ4BfEIhCdv1SLPvpUOBcUk9WeU9PuApTO3HRSvnE)8C7l1hx01J7(MP8yJNSvTE)SwNw01J7(MD4yJNSvC(iTvdsX)iTbnkWFFSBNw01J7PwRVF5ioTORh3JLVI8vtosZzDgND6EeZ4U(tAnUlGUhHJ7ogSEI7Gn70IEhUYxElT4Qp6(iJEhUk6iuhUgORpUO3HRI616W1aD9Xf9HKdF3FIEmtiODDZX1z7CYBYnYbjz2u6prE5VspakWp8j6Bt263SmZ(JZMMUE208WIztRE4ztBMnl5NIlH)JrgyN4725V8ciLCjd42jrkuUKtjulgnlGWSIzFYz)fehvcI4wAIjioDcYv7VGGEcqWh31)zgsy417pdrppi4JpggoE)zisjfmdV8dabU2O6pjoIS9(9x2qAcaZxy3Knhgz7h2FzR3bOy78lGKbkdx2o)nKjmSDUfoXy7Go9zCgbpJ8GTwzGmIUz)hroAgrIXDkgrsptaQhroIJOpGpIIwIpMe1rKLQB2u)midjqwh1c)bEmFqFRsBk5oLYRiN836RQEBY2oFY25JQCyn(rELo(xHPvT5J(oMy20rdh3ApTjkn7rhMwZrVwZrIwZ57SwdL)gQ1Ku7uNwtYFYCSQVZ2oxPFh29z1WSbwKxrjZVJp9bCn(bND7qZHo0Kunw3q7qgtYhqeJRohkT(uxkfdMCyw08naw2dhttsXGjFpgTuIakjyGsHfvXFUGaIgQIjG2MlG97gtUaAZjG1W1Ks(CbhLTZFhIdLfvGThnQbhl1ou97QQXlQgyY3yG40c2FN5VAfsBQMMsg3nRQiEJZcqJ8)yo4IFyw8nQTSovZzQjXQHrk4sdmkEnA0(IohpcO7BSaKVYHYsnPaRNQavmjfvhSlQcXjkTD(TMZsW9AmJMbjSPUdvrnm)wUi0tNOyPxDzRinAJVILeGeckVdaK7VUuQhF)2675X70ZJ3zN84D0wmyF94D07XJqY3bpE7tUhV4KSpQE82Y94Tv5XJjw6vxogKDuMhVqDsCp((ZR8qYXRRBbtsZZQ0L00vJNK2ukwy973pkqR3Lz1oiur4e9Q3264kpjGLrnaxpw6(BISjOWIA96pZ7djFLU(tnjLLPwphdcpeTEi5W2zRhEEqtTE2hvRx)1MOQFC(vdr0I1yJ7txNLuDQsYlwb6pW)YPA(TGD5MQXnzqk7fYtvIKRgXuLszqpJMxYs9OZXJXlbbHAlHQrLZiSebcoCgjSkt4OvyLnCTzN2Ks9rLNt)1aQYZr1QKjZls1ZyKhLoL)vJOsC)14PH)k39hwrGxXQ8XWJDvMas)orzaO2ngc1dkYcN3FXBel7AEzvtt6AujtH69ePT)IMiwMX8YiMkTgvIqO(grA7VoiTUAE1VxAAEkMl0p8OYok7TweLnYMDSSNXuMk9570X67zxvePDBhU68SuylPQaztx6FlubAJfUZ067nQtL6kznA)VF6kSiXdxx1TTIWOU(Dvs7MDkR6v3lmfwDBZot2sv7RkLEuDflvTVYu6r01SeX)QtHNYcYNxt63WYw5U9XQ)sMHjLm(eguWTdZQVeAy7(zdf9)YGPQ7)91bbsu7uTh11lNCJPdztsLzffLtg8Q21BKDaTt2PTRk82EIiTvGou63Yk6aOMiLFpRABGSzEd21RXcL39EtFrj4hORbR03ulmZVry9ve)QjIozdH3fgOyk)BvjpOsFxNqlZ2oPdSyBS9aUZ6yAEHMCndyek5VjpOwo7rDsMG8JT7V7UurMSJt1Y2ZmXrZN7x946Q9EC13yJTfW73WY2S9GJFGmQxyl3GaBZCBOMIk51ILyKBn)6EJWuphtAe05TEY0pyxToVG6S)hYRM9gPAzZ6E9GvprKQ5NjIfiqSFJHiPMW2FvJYvwNZP77JuvYIEQHOUVHyOBkzhRudgYNrm0DcSdVg27rEV20tjr15DGZNVcDdH0F0dOgOADeXePBS1Ekdu)90Al)zlG8U7cD973nbKne89AKuhZLuuVIBPl0SPqGoy39JeKSb7nAgSxFGs6HwCGFb0ShjloRkex8Bwg9YQQuxnxhXSrvFSSQVn2N1S2vSrctX9Ci5Hy)OPXCDP8HTNzQILDjCQiQBrfBUv73wTEnEkSaskZsZLjL6nmatzXN6ENY12VOXHZ7ZKWEXbk2YXWTCgICRCnBB4DFa4qxSePJfMCTCVEprxedCu0MYwyvk2znlneqMEJBUrmV6l7ppNXS0Xpy7pfQ2pGCZqoPRmlhi6Maq)4ZZfol7tqFN4W)TXx90Yoa2uNGdECSPS1RLiS0u2Y956x2dZi3(D)UPqB1efANyNOLVBBmptjDK5AzjXdzSroiudVxzXdPzZM(7Lp6f7LerV6S))]] )
+-- spec:RegisterPack( "Demonology", 20250818.1, [[Hekili:v31EZTnYr(plQszAblBkcirVYBjQR24DRAZMu(2l01LRUuhbbbaLqeianEiALsf)SF9m41my65bFzN8p2san(190VNhcyM9SppBAGxr4Sp5mYz8OBSVzOJZvFW2z20INxhoB6Ap)h9Uh(HeVvW)(ZHRstsJtV)zYTEoo1lGarEAzMpC7ztxugfx8NsMTahxBG21H(WLF)7Nn9HOGGWkAdZ9NnLq77gDZ7SV5h3o)Vg(uuEyW25RZIsZIkEE78fE0lKMSDEE0QTZlcZlIsUFiilzPlJIbj4p8h2o)HII15)4LxEFuXdLlg6NU6sG8YyVIO0e)mVLfKF3)YfXPlUS4HWnEzBasJsU8N8jK871m8VeLxKFzq4sVY4IlbQIt9F0nOvfmKGY2FJWYPT4)rc(BN)X0vRIG)F6Zj(WGjC8h(W4RDQi(NbfdCTUr72FR6g)(F7D2)yN8VzZMHzErblslYBgfWpFzw460SIldEQ8)5F8Rb5phz)Z3x8LV85Fk53wh1I0yZrAX4h)V(50)8p9X)50S0Sh(Lp)3St)YNQqA78F9N(O5qL)lF8pB)l)ZFDC6)7)4pEtXvp58BF4lbKbOhv3MpCDwi8yl8kUyYLldJDdsxfLqvDVnA5KIOvH3nAWzRdlg6fh9u4GZwuUC5W7ZIwLgLf6MU0n3ZplAzKF4WY144MxUcmsUaiem3bWiJymaFYllYBrC4BjHbtwgLLx4w8CMxc8FGe)201tYbM9KxCz4e7XheqEbb1avbqN4cQR7l9Ycgg(1q)YIWQhHO18IdtkWimHWSGJR8uRDFcc4wgfMivCeP7Okn37hmCL3xnz8)YlNsrkVCbQcIMPiYVMCoT0fnu(Gxq6g3fPX83)GKhiEQXtCKPafL4MUomjmRNRmE8160nHzU5rRFOkQLgvTjkoWnA168H5fqvJ7K8SuTczatEqOWrSlOeYcYhcjAJ8JkUZEW5lsZZhqbTrh6NMfoeuvjV8YzsUB5Aljzd6uYnzA)LkDnumj8PWmO0cvPphkgqvwPzBNhLd)4tErXeL0WwKb88(kyKazG9ARtBYGrs20ZYti7PqwY99IJDR(v3yOmtLriZZpYloxjkV8cD4Vm6(hkCZcx5fLKFRJZBlZjPZw6crfmwnPSkQiCvE)hIOz(7)A4JrXr)FBN)Fs9h2opjfQJLeggquxrqT3vP5Wvafmer1veMZttkFRDYQZ0FBJt4qbh5(WfL8u6JHUHFTimlXl2Ly)RGSYzmkzzzETjOfvWBeIcZx7Tjza2fVLglwNi5nV)cFVCwU3PlG67RJHYeevau6pYxy4IgZlvyOM3SqVaiuj(XWSAJ7GZptsIQxEXpnnMeaiMdR2lWQ7PfseY84Ijj7EET6Oone4hIf0p5dVZ5nqW3teFQH2dFWljGuM9(Y4aVelwhtPASQqBgTtoeBpOoJjpGdjIcKikXp82AcO(D8pDhrdoN2GH9ijI)p8o5Y(3xJdRMtGkIwRVi0l7rtQcO7KbTKIORAOJu6qWw1oSmNH3nXzmRxJzpfiMxThIPiH7IGs8kmJ2BNWZ3UAA(p7hdvh8YUpSiFIDN7mBLQ6qLRE351UebPPRgeewrj8Z0kDwVXzqx9XBNC1agQf4oGDErwyY9fpOS(XGQ)ZfsdVkkm)oUAfla7BUpmzU7dHYLRwuRav9i8bjeQByRxwbJ8ddx2hRx3hq2v(gpOQIAbNOmWeJlAuFT9Wh4syyEtxDw7lhBmZVhHTo6zBpD8TJ1)ihQKEdIKE9UXwcJsBkHAwCauogMnp0hfKrTVJ1vdajg8IskHeWB244E9A)b4jDRNwgxmcee7xaZdSlDCR8bTaIaDFZ0GInPU0epewb)okwCTfeZQoYkRNwQXb6mbRxJ0FAtuVYWyXgJvJ1zsaRFh6iAyMjpt7VAzDJvIsWvQd4LgKVlJiQ9MRKdBx7sKZM257Au7)U2CMtAt9js7AlZanZ825dqw6iOraAV(qF)EqkZxdxSKUYsLjfrXBNVbUnyw3o)HWmMw)hsBAEcF4zrwuYJG3OnTJ086q06RsBIi3L0Mn5UW0qBVEu(WLLjpKsUhybZ7XfH0an8XbLpos4JZEZhBi22pUmiuyebiMvU4z3npegVgSOHXXMcQJeqDoeqTDx5Lu6fJjO5RFEfKcaQJ5UjCrp9peXhMTeYF56LdZLb8M9F2npmRCLPS2bN1oQyTZXH1v(BUbLzvRDw)b)6Su)HEjp7ga5zBOckii1b8n2JT6D)8O7tawbb7qWIlDDyF(noJSmvgDukJo6KrrNxEz05yiJ2U5pN4xxGnAzZ6Eu9)UHX5HtgnC8B9ttcIi41nFUEr(dy0DTvpBguV6vAkO2q4e99G2HPcgoXCnGZbRbC6RbC(wRbqyObAGaVvEqNUndJMnAOVQWHvvyZOiotQVG4DA0rmoVXpfFht8g8R6e4tGKktqF5f5MzL2zjpwBwaR3CU9WXx0bbPYfTFGOKciTpKueibbfIxQL1DNFUYanjduzS32q2BxZE4PVGvaiwT3XBtTEL9OrwwiEFK0yKFMFD3EBECArJP1MBjJ07w1u8BW5NjFkFm6RMjIFheLDg61TgCUIfFS3sGWtfIptJN6eNbMnLI7CQ39h1dfgpVMPazg(xmMDDU68UQBmHrRu5z2GkBb8(Sw142gnxqLrZc5b75UE3eUv6DV8PCq9PCuKQAh8PCK4t58nWNY(46tzFC9PSr8PSL4tjYA1XsYteG5t5C89PO5P2L6FDjQ60usQ)oXwvWMfxJwD(ykSWst19Yl01d(wW)qw8HL2Urewle0UpnqN6GRtnjqvVo1rLZMfxRB7Qo1XyDQ93eDkTvj(PFPEjrr2rTj3mQArkyWP6io0U2lJG8Jsda60nitfuF3HitE1cDN)MaQzS(HQVVA5qEqMkbbt(TBAAu6qBhnFYNG8UV7OGPsbZPSBvuwwAgzwKlZaklZcdClsxLcxDdDjW50hkPw0Rhlg05qWXwXWzhssRpPH6eXgkfAsRPxkuN6Yy)kQmrEgxYsywz3pljekkgtAU3BDoz1JH75hMDqGkbtIAqvBqyEU2qNf7sv2AL6)6i(QCGUZSIawsRn)VIdtrp09FyA3mm7Vk3FkTEjR5pgcdRoDftuVZ6IBzF9YXJTbRdoNEVU9ZUCTqd7T30QMAXDXM8uTQbPBY9DxnQdCbQSWgRxiz3Wz3tIXVRgsYXnKCos8R3xhu4W2Q7o0Mmw2J1TDf92Td173E)tDylG47LTbsYFTBJRqVVMnWzq9IHWSnlYfYE7gZoRzu44rIbBCvRJ(iBe1EnKWWQ)zJRfmUDTQsaf2Okvco7gDzS82g7lCqOUT5O3uHIKKi9o5s6mcYfF2JQB95rBc5CaeM9yuY9soMznucmArCAAG7YYSNXUlm6cPuGDts(38csszIVVywVp)qyEy3w2vDK06LbSrp2FB6sc)ARw1pGF)0nl)C7CrKWqHfRvYrQCKYgyvDWZ(mLWTZBxw38TZxMghNUbSm0JeVF7XXR6uY3DeZLPMeL6Ut(fYPoa1L5IMZ9R0Zofw4s5Adx9MBTVc9W7nz0bpMWG9U)JEv86pkXEObcvjpLYM0cOgiJy1MpyzDuBoLAtxdM6Bzro3wdEh0uU5UlHzjv3fZEe5YTue8jci36CXC80gJULmXd0ZnHLGNRG3stw73CJrhJTwQNn9j46GMP9VCMXZMUXllbIPZNn9ZKdOmygsZkiX(zBN)AARVVMCEC)sjyxH8e5PKZYSxjmvsp6zC2hIgVpmh08)LOe4wK)it(yAcWj6TFDxVRSNIaaZIuM7wiC7ZT)QLbG2FxCragHeoWDKaUJAjw42gcQwjgLeoWVsL6O)rDaxJGqfhlUwL8RLfYOIJfJvnk4xwi8XGanMdV8LacNvkPNJTVxLItZOcNgZHFhgv6PNJT)GuLj65SarjkJUJmBiXIU7o8Adk1Ngrj4yNKejo14uYXQBK6eyMAsoDhz2O0AihETwd9PivcUrwdvuYXQpi1WlCccqm5y0CsH)Kb83j5gMhuff5v)fe9A5RjTk8pTNHimoozKvZaFpfzdfJDrI1jWDTKnsAeME7lon8T8DSX)0H83njxVRVWPNdJbNyxF8ZpNuxFZezdfJDrI1jWDMbzthP6mAHvjP6Y8UjkMudgk24OiFIg7NZ2PgFtqw8yZHankrMk7gZbPZivg27B5Ttn(MGSbAfjezQSBmh2d9(XjQzxIGD0Ko(eDYyrtik55KD2u75EB6PIfTnejJsPS2wlRXprSf9osSf8Nj2gR23qT(LF)06IS(BSw)YATEx4G8LLQDhKWsvXCp(WRVv4jE6buvjRJKtj6MGl7X9us5lMBFkL2tl6D4kFvjL7r4OWJ4BfEIhCdv1SLPvpUOBcUk9WeU9PuApTO3HRSvnE)8C7l1hx01J7(MP8yJNSvTE)SwNw01J7(MD4yJNSvC(iTvdsX)iTbnkWFFSBNw01J7PwRVF5ioTORh3JLVI8vtosZzDgND6EeZ4U(tAnUlGUhHJ7ogSEI7Gn70IEhUYxElT4Qp6(iJEhUk6iuhUgORpUO3HRI616W1aD9Xf9HKdF3FIEmtiODDZX1z7CYBYnYbjz2u6prE5VspakWp8j6Bt263SmZ(JZMMUE208WIztRE4ztBMnl5NIlH)JrgyN4725V8ciLCjd42jrkuUKtjulgnlGWSIzFYz)fehvcI4wAIjioDcYv7VGGEcqWh31)zgsy417pdrppi4JpggoE)zisjfmdV8dabU2O6pjoIS9(9x2qAcaZxy3Knhgz7h2FzR3bOy78lGKbkdx2o)nKjmSDUfoXy7Go9zCgbpJ8GTwzGmIUz)hroAgrIXDkgrsptaQhroIJOpGpIIwIpMe1rKLQB2u)midjqwh1c)bEmFqFRsBk5oLYRiN836RQEBY2oFY25JQCyn(rELo(xHPvT5J(oMy20rdh3ApTjkn7rhMwZrVwZrIwZ57SwdL)gQ1Ku7uNwtYFYCSQVZ2oxPFh29z1WSbwKxrjZVJp9bCn(bND7qZHo0Kunw3q7qgtYhqeJRohkT(uxkfdMCyw08naw2dhttsXGjFpgTuIakjyGsHfvXFUGaIgQIjG2MlG97gtUaAZjG1W1Ks(CbhLTZFhIdLfvGThnQbhl1ou97QQXlQgyY3yG40c2FN5VAfsBQMMsg3nRQiEJZcqJ8)yo4IFyw8nQTSovZzQjXQHrk4sdmkEnA0(IohpcO7BSaKVYHYsnPaRNQavmjfvhSlQcXjkTD(TMZsW9AmJMbjSPUdvrnm)wUi0tNOyPxDzRinAJVILeGeckVdaK7VUuQhF)2675X70ZJ3zN84D0wmyF94D07XJqY3bpE7tUhV4KSpQE82Y94Tv5XJjw6vxogKDuMhVqDsCp((ZR8qYXRRBbtsZZQ0L00vJNK2ukwy973pkqR3Lz1oiur4e9Q3264kpjGLrnaxpw6(BISjOWIA96pZ7djFLU(tnjLLPwphdcpeTEi5W2zRhEEqtTE2hvRx)1MOQFC(vdr0I1yJ7txNLuDQsYlwb6pW)YPA(TGD5MQXnzqk7fYtvIKRgXuLszqpJMxYs9OZXJXlbbHAlHQrLZiSebcoCgjSkt4OvyLnCTzN2Ks9rLNt)1aQYZr1QKjZls1ZyKhLoL)vJOsC)14PH)k39hwrGxXQ8XWJDvMas)orzaO2ngc1dkYcN3FXBel7AEzvtt6AujtH69ePT)IMiwMX8YiMkTgvIqO(grA7VoiTUAE1VxAAEkMl0p8OYok7TweLnYMDSSNXuMk9570X67zxvePDBhU68SuylPQaztx6FlubAJfUZ067nQtL6kznA)VF6kSiXdxx1TTIWOU(Dvs7MDkR6v3lmfwDBZot2sv7RkLEuDflvTVYu6r01SeX)QtHNYcYNxt63WYw5U9XQ)sMHjLm(eguWTdZQVeAy7(zdf9)YGPQ7)91bbsu7uTh11lNCJPdztsLzffLtg8Q21BKDaTt2PTRk82EIiTvGou63Yk6aOMiLFpRABGSzEd21RXcL39EtFrj4hORbR03ulmZVry9ve)QjIozdH3fgOyk)BvjpOsFxNqlZ2oPdSyBS9aUZ6yAEHMCndyek5VjpOwo7rDsMG8JT7V7UurMSJt1Y2ZmXrZN7x946Q9EC13yJTfW73WY2S9GJFGmQxyl3GaBZCBOMIk51ILyKBn)6EJWuphtAe05TEY0pyxToVG6S)hYRM9gPAzZ6E9GvprKQ5NjIfiqSFJHiPMW2FvJYvwNZP77JuvYIEQHOUVHyOBkzhRudgYNrm0DcSdVg27rEV20tjr15DGZNVcDdH0F0dOgOADeXePBS1Ekdu)90Al)zlG8U7cD973nbKne89AKuhZLuuVIBPl0SPqGoy39JeKSb7nAgSxFGs6HwCGFb0ShjloRkex8Bwg9YQQuxnxhXSrvFSSQVn2N1S2vSrctX9Ci5Hy)OPXCDP8HTNzQILDjCQiQBrfBUv73wTEnEkSaskZsZLjL6nmatzXN6ENY12VOXHZ7ZKWEXbk2YXWTCgICRCnBB4DFa4qxSePJfMCTCVEprxedCu0MYwyvk2znlneqMEJBUrmV6l7ppNXS0Xpy7pfQ2pGCZqoPRmlhi6Maq)4ZZfol7tqFN4W)TXx90Yoa2uNGdECSPS1RLiS0u2Y956x2dZi3(D)UPqB1efANyNOLVBBmptjDK5AzjXdzSroiudVxzXdPzZM(7Lp6f7LerV6S))]] )
+
+spec:RegisterPriority( "Demo Test", 20250819.1,
+[[
+## https://github.com/simulationcraft/simc/blob/thewarwithin/ActionPriorityLists/default/warlock_demonology.simc
+## SimulationCraft Commit Sync: e599542
+## Date: 2025-08-18
+
+## PW-1: https://www.raidbots.com/simbot/report/dvuXjHdsyi1DgtqqTAnJpi
+## PW-5: https://www.raidbots.com/simbot/report/b5kQDoKACzSrorhETW1oqN
+##  HAC: https://www.raidbots.com/simbot/report/sECK1EzH5oZjB8t3v2J9qd
+
+actions.precombat+=/fel_domination,if=time>0&!pet.alive&!buff.grimoire_of_sacrifice.up
+actions.precombat+=/summon_pet,if=!pet.alive&!buff.grimoire_of_sacrifice.up
+## actions.precombat+=/variable,name=first_tyrant_time,op=set,value=15
+## actions.precombat+=/variable,name=first_tyrant_time,op=add,value=action.grimoire_felguard.execute_time,if=talent.grimoire_felguard.enabled
+## actions.precombat+=/variable,name=first_tyrant_time,op=add,value=action.summon_vilefiend.execute_time,if=talent.summon_vilefiend.enabled
+## actions.precombat+=/variable,name=first_tyrant_time,op=add,value=gcd.max,if=talent.grimoire_felguard.enabled|talent.summon_vilefiend.enabled
+## actions.precombat+=/variable,name=first_tyrant_time,op=sub,value=action.summon_demonic_tyrant.execute_time+action.shadow_bolt.execute_time
+## actions.precombat+=/variable,name=first_tyrant_time,op=min,value=10
+## actions.precombat+=/variable,name=in_opener,op=set,value=1
+actions.precombat+=/power_siphon,if=buff.wild_imps.stack>1
+actions.precombat+=/demonbolt,if=soul_shards.deficit>1&(boss&buff.demonic_core.down|!boss&buff.demonic_core.up)
+actions.precombat+=/shadow_bolt
+
+## Executed every time the actor is available.
+actions+=/axe_toss
+actions+=/potion,if=pet.demonic_tyrant.active
+actions+=/call_action_list,name=racials,if=pet.demonic_tyrant.active|boss&fight_remains<22,use_off_gcd=1
+actions+=/call_action_list,name=items,use_off_gcd=1
+## [Hekili] Opener not needed in most recent testing.
+## actions+=/call_action_list,name=opener,if=time<variable.first_tyrant_time
+## actions+=/invoke_external_buff,name=power_infusion,if=variable.imp_despawn&variable.imp_despawn<time+gcd.max*6+cast_time
+## [Hekili] Simplified logic.
+## actions+=/summon_demonic_tyrant,if=variable.imp_despawn&pet.dreadstalker.active&(!talent.summon_vilefiend|cooldown.summon_vilefiend.remains)&(!talent.grimoire_felguard|cooldown.grimoire_felguard.remains)&(variable.imp_despawn<time+gcd.max+cast_time|buff.wild_imps.stack>=9-2*prev_gcd.1.hand_of_guldan)
+actions+=/summon_demonic_tyrant,if=buff.dreadstalkers.up&action.hand_of_guldan.time_since<action.call_dreadstalkers.time_since&(time>10|buff.wild_imps.stack>7-prev_gcd.1.hand_of_guldan)&(!talent.summon_vilefiend|cooldown.summon_vilefiend.remains)&(!talent.grimoire_felguard|cooldown.grimoire_felguard.remains)
+actions+=/grimoire_felguard,if=cooldown.summon_demonic_tyrant.remains<=15&cooldown.call_dreadstalkers.remains<10
+actions+=/summon_vilefiend,if=cooldown.summon_demonic_tyrant.remains>=25+cast_time|cooldown.summon_demonic_tyrant.remains<=13&cooldown.call_dreadstalkers.remains<10
+actions+=/call_dreadstalkers,if=cooldown.summon_demonic_tyrant.remains>=10|cooldown.summon_demonic_tyrant.remains<=10
+actions+=/demonbolt,cycle_targets=1,if=buff.demonic_core.stack>=3-(talent.doom&debuff.doom.down)*2&soul_shard<=3&talent.doom
+actions+=/demonic_strength,if=pet.demonic_tyrant.active&active_enemies>1
+actions+=/bilescourge_bombers,if=active_enemies>1
+actions+=/hand_of_guldan,if=demonic_art&soul_shard>=3
+actions+=/variable,name=impl,op=set,value=buff.tyrant.down,if=active_enemies>1+(talent.sacrificed_souls.enabled)
+actions+=/variable,name=impl,op=set,value=buff.tyrant.remains<6,if=active_enemies>2+(talent.sacrificed_souls.enabled)&active_enemies<5+(talent.sacrificed_souls.enabled)
+actions+=/variable,name=impl,op=set,value=buff.tyrant.remains<8,if=active_enemies>4+(talent.sacrificed_souls.enabled)
+actions+=/implosion,if=cooldown.summon_demonic_tyrant.remains_expected>10&active_enemies>3&set_bonus.tww2_4pc&buff.wild_imps.stack>7&!buff.demonic_core.react&!prev_gcd.1.implosion|!set_bonus.tww2_4pc&active_enemies>2&two_cast_imps>2&!prev_gcd.1.implosion&variable.impl
+actions+=/ruination
+actions+=/demonbolt,cycle_targets=1,if=soul_shard<4&buff.demonic_core.stack>=3&talent.doom
+actions+=/demonbolt,if=soul_shard<4&buff.demonic_core.stack>=3&!talent.doom
+actions+=/power_siphon,if=!buff.demonic_core.up
+actions+=/infernal_bolt,if=soul_shard<3
+actions+=/hand_of_guldan,if=soul_shard>=3
+actions+=/demonbolt,if=soul_shard<4&buff.demonic_core.react
+actions+=/shadow_bolt
+actions+=/infernal_bolt
+
+## [Hekili] Variables moved from precombat since they aren't used until we get here.
+actions.items=variable,name=trinket_1_buffs,value=trinket.1.has_use_buff|trinket.1.is.funhouse_lens
+actions.items+=/variable,name=trinket_2_buffs,value=trinket.2.has_use_buff|trinket.2.is.funhouse_lens
+actions.items+=/variable,name=trinket_1_exclude,value=trinket.1.is.ruby_whelp_shell
+actions.items+=/variable,name=trinket_2_exclude,value=trinket.2.is.ruby_whelp_shell
+actions.items+=/variable,name=trinket_1_manual,value=trinket.1.is.spymasters_web|trinket.1.is.imperfect_ascendancy_serum
+actions.items+=/variable,name=trinket_2_manual,value=trinket.2.is.spymasters_web|trinket.2.is.imperfect_ascendancy_serum
+actions.items+=/variable,name=trinket_1_buff_duration,value=trinket.1.proc.any_dps.duration+(trinket.1.is.funhouse_lens*15)+(trinket.1.is.signet_of_the_priory*20)
+actions.items+=/variable,name=trinket_2_buff_duration,value=trinket.2.proc.any_dps.duration+(trinket.2.is.funhouse_lens*15)+(trinket.2.is.signet_of_the_priory*20)
+actions.items+=/variable,name=trinket_1_sync,op=setif,value=1,value_else=0.5,condition=variable.trinket_1_buffs&(trinket.1.cooldown.duration%%cooldown.summon_demonic_tyrant.duration=0|cooldown.summon_demonic_tyrant.duration%%trinket.1.cooldown.duration=0)
+actions.items+=/variable,name=trinket_2_sync,op=setif,value=1,value_else=0.5,condition=variable.trinket_2_buffs&(trinket.2.cooldown.duration%%cooldown.summon_demonic_tyrant.duration=0|cooldown.summon_demonic_tyrant.duration%%trinket.2.cooldown.duration=0)
+actions.items+=/variable,name=damage_trinket_priority,op=setif,value=2,value_else=1,condition=!variable.trinket_1_buffs&!variable.trinket_2_buffs&trinket.2.ilvl>trinket.1.ilvl
+actions.items+=/variable,name=trinket_priority,op=setif,value=2,value_else=1,condition=!variable.trinket_1_buffs&variable.trinket_2_buffs|variable.trinket_2_buffs&((trinket.2.cooldown.duration%variable.trinket_2_buff_duration)*(1.5+trinket.2.has_buff.intellect)*(variable.trinket_2_sync))>(((trinket.1.cooldown.duration%variable.trinket_1_buff_duration)*(1.5+trinket.1.has_buff.intellect)*(variable.trinket_1_sync))*(1+((trinket.1.ilvl-trinket.2.ilvl)%100)))
+actions.items+=/use_item,use_off_gcd=1,slot=trinket1,if=variable.trinket_1_buffs&!variable.trinket_1_manual&(!pet.demonic_tyrant.active&trinket.1.cast_time>0|!trinket.1.cast_time>0)&(pet.demonic_tyrant.active|!talent.summon_demonic_tyrant|variable.trinket_priority=2&cooldown.summon_demonic_tyrant.remains>20&!pet.demonic_tyrant.active&trinket.2.cooldown.remains<cooldown.summon_demonic_tyrant.remains+5)&(variable.trinket_2_exclude|!trinket.2.has_cooldown|trinket.2.cooldown.remains|variable.trinket_priority=1&!variable.trinket_2_manual)|variable.trinket_1_buff_duration>=fight_remains
+actions.items+=/use_item,use_off_gcd=1,slot=trinket2,if=variable.trinket_2_buffs&!variable.trinket_2_manual&(!pet.demonic_tyrant.active&trinket.2.cast_time>0|!trinket.2.cast_time>0)&(pet.demonic_tyrant.active|!talent.summon_demonic_tyrant|variable.trinket_priority=1&cooldown.summon_demonic_tyrant.remains>20&!pet.demonic_tyrant.active&trinket.1.cooldown.remains<cooldown.summon_demonic_tyrant.remains+5)&(variable.trinket_1_exclude|!trinket.1.has_cooldown|trinket.1.cooldown.remains|variable.trinket_priority=2&!variable.trinket_1_manual)|variable.trinket_2_buff_duration>=fight_remains
+actions.items+=/use_item,use_off_gcd=1,slot=trinket1,if=!variable.trinket_1_buffs&!variable.trinket_1_manual&((variable.damage_trinket_priority=1|trinket.2.cooldown.remains)&(trinket.1.cast_time>0&!pet.demonic_tyrant.active|!trinket.1.cast_time>0)|(time<20&variable.trinket_2_buffs)|cooldown.summon_demonic_tyrant.remains_expected>20)
+actions.items+=/use_item,use_off_gcd=1,slot=trinket2,if=!variable.trinket_2_buffs&!variable.trinket_2_manual&((variable.damage_trinket_priority=2|trinket.1.cooldown.remains)&(trinket.2.cast_time>0&!pet.demonic_tyrant.active|!trinket.2.cast_time>0)|(time<20&variable.trinket_1_buffs)|cooldown.summon_demonic_tyrant.remains_expected>20)
+actions.items+=/use_item,use_off_gcd=1,name=spymasters_web,if=pet.demonic_tyrant.active&boss&fight_remains<=80&buff.spymasters_report.stack>=30&(!variable.trinket_1_buffs&trinket.2.is.spymasters_web|!variable.trinket_2_buffs&trinket.1.is.spymasters_web)|boss&fight_remains<=20&(trinket.1.cooldown.remains&trinket.2.is.spymasters_web|trinket.2.cooldown.remains&trinket.1.is.spymasters_web|!variable.trinket_1_buffs|!variable.trinket_2_buffs)
+actions.items+=/use_item,use_off_gcd=1,name=imperfect_ascendancy_serum,if=pet.demonic_tyrant.active|boss&fight_remains<=30
+actions.items+=/use_item,name=mirror_of_fractured_tomorrows,if=trinket.1.is.mirror_of_fractured_tomorrows&variable.trinket_priority=2|trinket.2.is.mirror_of_fractured_tomorrows&variable.trinket_priority=1
+actions.items+=/use_item,slot=trinket1,if=!variable.trinket_1_buffs&(variable.damage_trinket_priority=1|trinket.2.cooldown.remains)
+actions.items+=/use_item,slot=trinket2,if=!variable.trinket_2_buffs&(variable.damage_trinket_priority=2|trinket.1.cooldown.remains)
+actions.items+=/use_item,use_off_gcd=1,slot=main_hand,name=!neural_synapse_enhancer
+actions.items+=/use_item,use_off_gcd=1,slot=main_hand,name=neural_synapse_enhancer,if=(pet.demonic_tyrant.active|boss&fight_remains<=15)&!variable.trinket_1_buffs&!variable.trinket_2_buffs
+actions.items+=/use_item,use_off_gcd=1,slot=main_hand,name=neural_synapse_enhancer,if=(pet.demonic_tyrant.active|boss&fight_remains<=15|trinket.2.cooldown.remains>cooldown.summon_demonic_tyrant.remains)&variable.trinket_2_buffs
+actions.items+=/use_item,use_off_gcd=1,slot=main_hand,name=neural_synapse_enhancer,if=(pet.demonic_tyrant.active|boss&fight_remains<=15|trinket.1.cooldown.remains>cooldown.summon_demonic_tyrant.remains)&variable.trinket_1_buffs
+
+## [Hekili] Not used.
+## actions.opener=summon_demonic_tyrant,if=buff.wild_imps.stack>=9&buff.dreadstalkers.remains&(buff.vilefiend.up|!talent.summon_vilefiend)&(buff.grimoire_felguard.up|cooldown.grimoire_felguard.remains>30|!talent.grimoire_felguard)
+## actions.opener+=/grimoire_felguard,if=soul_shard>=5-talent.fel_invocation
+## actions.opener+=/summon_vilefiend,if=soul_shard=5
+## actions.opener+=/shadow_bolt,if=soul_shard<5&cooldown.call_dreadstalkers.up
+## actions.opener+=/call_dreadstalkers,if=soul_shard=5
+## actions.opener+=/Ruination
+## actions.opener+=/hand_of_guldan,if=soul_shard>=3&buff.infernal_bolt.up
+## actions.opener+=/infernal_bolt
+## actions.opener+=/shadow_bolt,if=soul_shard<5&buff.dreadstalkers.remains&pet.wild_imp.active<3
+## actions.opener+=/hand_of_guldan,if=soul_shard>=3&pet.wild_imp.active<10
+## actions.opener+=/demonbolt,if=buff.demonic_core.react&buff.dreadstalkers.remains&soul_shard<4
+## actions.opener+=/hand_of_guldan,if=variable.first_tyrant_time<gcd.max+action.summon_demonic_tyrant.cast_time
+## actions.opener+=/shadow_bolt,if=buff.dreadstalkers.remains
+
+actions.racials=berserking,use_off_gcd=1
+actions.racials+=/blood_fury
+actions.racials+=/fireblood
+actions.racials+=/ancestral_call
+
+## [Hekili] These aren't needed.
+## actions.variables=variable,name=next_tyrant_cd,op=set,value=cooldown.summon_demonic_tyrant.remains_expected
+## actions.variables+=/variable,name=in_opener,op=set,value=0,if=pet.demonic_tyrant.active
+## [Hekili] Simplified Tyrant conditions following simc testing on 2025-08-18.
+## actions.variables+=/variable,name=imp_despawn,op=set,value=buff.dreadstalkers.remains+time,if=prev_gcd.1.hand_of_guldan&buff.dreadstalkers.up&cooldown.summon_demonic_tyrant.remains<13&variable.imp_despawn=0
+## actions.variables+=/variable,name=imp_despawn,op=set,value=variable.imp_despawn>?buff.vilefiend.remains+time,if=variable.imp_despawn&buff.vilefiend.up
+## actions.variables+=/variable,name=imp_despawn,op=set,value=variable.imp_despawn>?buff.grimoire_felguard.remains+time,if=variable.imp_despawn&buff.grimoire_felguard.up
+## actions.variables+=/variable,name=imp_despawn,op=set,value=0,if=buff.tyrant.up
+## [Hekili] Not used.
+## actions.variables+=/variable,name=pool_cores_for_tyrant,op=set,value=cooldown.summon_demonic_tyrant.remains<20&variable.next_tyrant_cd<20&(buff.demonic_core.stack<=2|!buff.demonic_core.up)&cooldown.summon_vilefiend.remains<gcd.max*8&cooldown.call_dreadstalkers.remains<gcd.max*8
+]],
+{
+    desc = [[20250819: Removed opener and revised Implosion.]]
+} )
